refactor: move Hammerstein kernel estimation and response prediction logic into core library functions
